--- a/test.js
+++ b/test.js
@@ -1,0 +1,9 @@
+const transform = require('./dist/camaro')
+
+console.log(transform);
+
+;(async function () {
+    const xml = '<root>world</root>'
+    const result = await transform(xml, { hello: '/root'})
+    console.log(result);
+})()