… dist/camaro.js
… dist/camaro.wasm